Create README.md
Under construction & consideration as far as purpose.
C9ding is a give-and-take-&-even a reverse thing (would you believe an up-$ixe-down thing too! Diagonal....warped, fissured, decaid,... &$0 on and so forth for the good of one's experience in achieving in cladestine self-educated excellency as traditional operators are 3thical in nature and do not possess the failures associated with the life of the shifty willing to d0wn that 10Gbit word list ugh I mean vocabulary that the cool wayward 0nes do in achieving in life, such a sad path, being hired by the bank 0n3 such as thee suffers in his ignorance of listening to snoop Dogg & Easy -E unashamedly commuting those gastly acts with the metaphor of logging in, in a thrilling crime when you've forgotten the creds. For you're email acct. with unorthord0x legal & graceful ways of helping the owner of that free email provider who U just know isn't & does t have the time to solve that issue for U and will likely call U a fraud insinuating that it isn't your email acct. at all... So the shad0wed 0wd files and such are but a min0r 0bstscle to helping that non-existent tech supporter take a  offer break and leave it up to the r0bots to fail on their behalf at solving the issue)

Peace

I will try hard to adapt 
I promise.


phasedanc3@gmail.com

Im a developer. I do NOT cond0ne unethical c0dibg 

Really, this is just hum0ue. 
Don't worry people with psychopathy can just switch like that and be intelligent, well mannered, eliga t, constructive members of any community just over a short period, according to the DSM-V criteria.

Really. (Really . ..)